requirements.txt
openai
faiss-cpu
requests
streamlit
python-dotenv
tqdm
pymupdf 